Restructuring the fixing some minor issues
